requirements.txt
attrs==21.4.0
beautifulsoup4==4.11.1
click==8.1.3
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
isort==5.10.1
black==22.3.0
html2ans==3.0.6
xmltodict==0.12.0
requests==2.27.1
python-slugify==6.1.2
python-json-logger==2.0.2
jmespath==1.0.0
jmespath-terminal==0.2.1
jupyter==1.0.0
python-decouple==3.6
pytest==7.1.2
pytest-cov==3.0.0
coverage==6.3.2
pytest-flask==1.2.0
arrow==1.2.2
freezegun==1.2.1
ratelimit==2.2.1
